+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdErrorSpec.java
@@ -8,9 +8,6 @@ import com.intellij.psi.PsiElement;
 public interface FsdErrorSpec extends PsiElement {
 
   @NotNull
-  List<FsdAttributeList> getAttributeListList();
-
-  @NotNull
   PsiElement getIdentifier();
 
 }