+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdTypes.java
@@ -12,6 +12,10 @@ public interface FsdTypes {
   IElementType ATTRIBUTE_LIST = new FsdElementType("ATTRIBUTE_LIST");
   IElementType ATTRIBUTE_PARAMETER = new FsdElementType("ATTRIBUTE_PARAMETER");
   IElementType DATA_SPEC = new FsdElementType("DATA_SPEC");
+  IElementType DECORATED_ENUM_VALUE = new FsdElementType("DECORATED_ENUM_VALUE");
+  IElementType DECORATED_ERROR_SPEC = new FsdElementType("DECORATED_ERROR_SPEC");
+  IElementType DECORATED_FIELD = new FsdElementType("DECORATED_FIELD");
+  IElementType DECORATED_SERVICE_ITEM = new FsdElementType("DECORATED_SERVICE_ITEM");
   IElementType ENUM_SPEC = new FsdElementType("ENUM_SPEC");
   IElementType ENUM_VALUE = new FsdElementType("ENUM_VALUE");
   IElementType ENUM_VALUE_LIST = new FsdElementType("ENUM_VALUE_LIST");
@@ -78,6 +82,18 @@ public interface FsdTypes {
       }
       else if (type == DATA_SPEC) {
         return new FsdDataSpecImpl(node);
+      }
+      else if (type == DECORATED_ENUM_VALUE) {
+        return new FsdDecoratedEnumValueImpl(node);
+      }
+      else if (type == DECORATED_ERROR_SPEC) {
+        return new FsdDecoratedErrorSpecImpl(node);
+      }
+      else if (type == DECORATED_FIELD) {
+        return new FsdDecoratedFieldImpl(node);
+      }
+      else if (type == DECORATED_SERVICE_ITEM) {
+        return new FsdDecoratedServiceItemImpl(node);
       }
       else if (type == ENUM_SPEC) {
         return new FsdEnumSpecImpl(node);