+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdVisitor.java
@@ -23,6 +23,22 @@ public class FsdVisitor extends PsiElementVisitor {
     visitPsiElement(o);
   }
 
+  public void visitDecoratedEnumValue(@NotNull FsdDecoratedEnumValue o) {
+    visitPsiElement(o);
+  }
+
+  public void visitDecoratedErrorSpec(@NotNull FsdDecoratedErrorSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitDecoratedField(@NotNull FsdDecoratedField o) {
+    visitPsiElement(o);
+  }
+
+  public void visitDecoratedServiceItem(@NotNull FsdDecoratedServiceItem o) {
+    visitPsiElement(o);
+  }
+
   public void visitEnumSpec(@NotNull FsdEnumSpec o) {
     visitPsiElement(o);
   }